YUK-9773 CBC 8000- Support Float point data objects Added serializeValue<T>() and restoreValue<T>() functions to DNPObject base class, now uses int32_t and int16_t in AnalogOutput and AnalogInput serialization code

+++ yukon-server/Protocol/include/dnp_object_analoginput.h
@@ -61,7 +61,7 @@ public:
 
     virtual CtiPointDataMsg *getPoint( const TimeCTO *cto ) const;
 
-    void setValue(long value);
+    void setValue(double value);
     void setOnlineFlag(bool online);
 };
